requirements.txt
absl-py==1.4.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
imageio==2.25.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2023.2.3
torch @ file:///home/huyu/dev/pkg/whl/torch-1.13.0%2Bcu116-cp310-cp310-linux_x86_64.whl
torchvision @ file:///home/huyu/dev/pkg/whl/torchvision-0.14.0%2Bcu116-cp310-cp310-linux_x86_64.whl
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
